05 detail

+++ lib/data/model/product.dart
@@ -103,7 +103,7 @@ class ShProduct {
   String? average_rating;
 
   // ignore: non_constant_identifier_names
-  int? rating_count;
+  double? rating_count;
 
   // ignore: non_constant_identifier_names
   List<dynamic>? related_ids;
